move QueryRow
