+++ sharding-core/src/main/java/io/shardingsphere/core/executor/sql/execute/result/AggregationDistinctQueryResult.java
@@ -23,6 +23,7 @@ import com.google.common.collect.Iterators;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Multimap;
 import io.shardingsphere.core.constant.AggregationType;
+import io.shardingsphere.core.executor.sql.execute.result.row.QueryRow;
 import io.shardingsphere.core.merger.QueryResult;
 import io.shardingsphere.core.parsing.parser.context.selectitem.AggregationDistinctSelectItem;
 import io.shardingsphere.core.parsing.parser.context.selectitem.AggregationSelectItem;
